+++ packages/internal-test/web/src/app/scrypt-benchmark/scrypt-benchmark.page.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ABFA, Err, Result } from '@soulwallet_test/keyvault';
+import { ABFA, Err, Result } from '@soulwallet/keyvault';
 
 @Component({
   selector: 'app-scrypt-benchmark',