+++ packages/soulwallet-decoder/src/main.ts
@@ -1,7 +1,7 @@
-import { TokenInfo, AddressType } from '@soulwallet_test/assets';
+import { TokenInfo, AddressType } from '@soulwallet/assets';
 import { DecodeUserOp } from './decoder.js';
 import { DecodeResult, Method } from './interface/decodeData.js';
-import { Ok, Err, Result } from '@soulwallet_test/result';
+import { Ok, Err, Result } from '@soulwallet/result';
 
 export {
     DecodeUserOp,