+++ packages/soulwallet-sdk/src/interface/ISoulWallet.ts
@@ -1,6 +1,6 @@
 import { UserOperation } from "./UserOperation.js";
 import { UserOpErrors } from "./IUserOpErrors.js";
-import { Result } from '@soulwallet_test/result';
+import { Result } from '@soulwallet/result';
 import { ECCPoint } from "../tools/webauthn.js";
 
 /**