+++ packages/soulwallet-keyvault/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @soulwallet/keyvault
 
+## 0.1.3
+
+### Patch Changes
+
+- passkey version merged into soulwallet@npmjs.com
+- Updated dependencies
+  - @soulwallet/result@0.1.2
+
 ## 0.1.2
 
 ### Patch Changes